+++ android/app/src/main/java/com/dklee/squadidle/MainActivity.java
@@ -1,0 +1,5 @@
+package com.dklee.squadidle;
+
+import com.getcapacitor.BridgeActivity;
+
+public class MainActivity extends BridgeActivity {}